aider: Added streamlit and pandas to requirements.txt.
# Aider chat conversation:

USER: create a requirements.txt file with all the needed dependencies.  Automatically add to that file any new depencies

ASSISTANT: To create a `requirements.txt` file, we will need to list all the dependencies used in the `fx.py` file. From the code, I can see that the dependencies are `streamlit` and `pandas`.

Here is the `requirements.txt` file:

```python
new_file/requirements.txt
<<<<<<< HEAD
=======
streamlit
pandas
>>>>>>> updated
```

To automatically add new dependencies, you would need to manually add them to the `requirements.txt` file whenever you import a new library in your Python files. There isn't a built-in way to automatically update the `requirements.txt` file when a new import statement is added to a Python file.

However, you can use pip freeze to generate a `requirements.txt` file with all the packages installed in your current environment. This can be done by running the command `pip freeze > requirements.txt` in your terminal. This will create a `requirements.txt` file with all the installed packages, but it might include packages that are not used in your project.

Please note that this command should be run in the same environment where you run your Python files. If you are using a virtual environment, make sure to activate it before running the command.